feat: Refactored Table.java to TableBuilder.java
